Datatype fix for question type

+++ models/Test.cs
@@ -7,9 +7,9 @@ namespace models
         public Guid Id { get; set; }
         public string Text { get; set; }
         public string Description { get; set; }
-        public Int16 Sequence { get; set; }
+        public int Sequence { get; set; }
         public string Subject { get; set; }
-        public Int16 DurationMinutes { get; set; }
+        public int DurationMinutes { get; set; }
 
         public DateTime CreatedAt { get; set; }
         public DateTime ModifiedAt { get; set; }
